// src/companies/custom-fields/enums/custom-field-type.enum.ts
export enum CustomFieldType {
  TEXT = 'text',           // Texto simple
  NUMBER = 'number',       // Números
  DATE = 'date',          // Fechas
  EMAIL = 'email',        // Correos electrónicos
  PHONE = 'phone',        // Teléfonos
  URL = 'url',            // URLs
  BOOLEAN = 'boolean',    // Verdadero/Falso
  SELECT = 'select',      // Selección única
  MULTISELECT = 'multiselect', // Selección múltiple
  SCHEDULE = 'schedule',  // Horarios
  ADDRESS = 'address',    // Direcciones
  FILE = 'file',         // Archivos/documentos
  JSON = 'json'          // Datos estructurados como JSON
}